+++ Applications/Pgan.PoracleWebNet.App/ClientApp/src/app/modules/invasions/invasion-edit-dialog.component.ts
@@ -12,7 +12,7 @@ import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatTabsModule } from '@angular/material/tabs';
 import { TranslateModule } from '@ngx-translate/core';
 
-import { EVENT_TYPE_INFO, getDisplayName, getGruntIconUrl, isEventType } from './invasion.constants';
+import { EVENT_TYPE_INFO, getDisplayName, getGruntIconUrl, hasNoGenderVariants, isEventType } from './invasion.constants';
 import { Invasion, InvasionUpdate } from '../../core/models';
 import { AuthService } from '../../core/services/auth.service';
 import { I18nService } from '../../core/services/i18n.service';
@@ -60,6 +60,7 @@ export class InvasionEditDialogComponent {
     template: [this.data.template ?? ''],
   });
 
+  readonly hideGender = hasNoGenderVariants(this.data.gruntType);
   readonly isEvent = isEventType(this.data.gruntType);
   readonly isWebhook = inject(AuthService).isImpersonating();
 
@@ -109,7 +110,7 @@ export class InvasionEditDialogComponent {
       .update(this.data.uid, {
         clean: v.clean ? 1 : 0,
         distance: dist,
-        gender: this.isEvent ? 0 : (v.gender ?? 0),
+        gender: this.hideGender ? 0 : (v.gender ?? 0),
         gruntType: this.data.gruntType ?? '',
         ping: v.ping || null,
         template: v.template || null,